[uni][vio] re-arrange routes
