+++ app/playground/components/color-mode-toggle.tsx
@@ -4,13 +4,14 @@ import * as React from "react"
 import {Moon, Sun} from "lucide-react"
 import {useTheme} from "next-themes"
 
-import {Button} from "@/components/ui/button"
+// import {Button} from "@/components/ui/button"
 import {
     DropdownMenu,
     DropdownMenuContent,
     DropdownMenuItem,
     DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
+import {Button} from "@/components/ui/button";
 
 export function ColorModeToggle() {
     const {setTheme} = useTheme()
