+++ tailwind.config.ts
@@ -8,7 +8,14 @@ const config: Config = {
     "./app/**/*.{js,ts,jsx,tsx,mdx}",
   ],
   theme: {
-  	extend: {
+	container: {
+	  center: true,
+	  padding: "2rem",
+	  screens: {
+		"2xl": "1400px"
+	  }
+	},
+	extend: {
   		colors: {
   			background: 'hsl(var(--background))',
   			foreground: 'hsl(var(--foreground))',
